requirements.txt
flask==3.1.2
flask-cors==6.0.1
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
requests==2.32.5
beautifulsoup4==4.14.2
trafilatura==1.6.3
numpy==2.3.5
scikit-learn==1.7.2
groq==0.36.0
apscheduler==3.11.1
langdetect==1.0.9
gunicorn==23.0.0
